backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pandas>=2.0.0
numpy>=1.24.2
scikit-learn>=1.2.2
pandas-datareader>=0.10.0
joblib>=1.3.0
aiohttp>=3.8.0
web3>=6.0.0
eth-account>=0.8.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.28.2
matplotlib>=3.7.1
seaborn>=0.12.2 